Sets up StyleData Context
