+++ src/components/overview/Overview.jsx
@@ -1,19 +1,23 @@
-import React, { useState, useContext } from 'react';
+import React, { useState, useContext, useEffect } from 'react';
+import axios from 'axios';
 import styles from './Overview.css';
 import withProductDetails from './ProductDetailWrapper';
-import withStyleDetails from './StyleDetailWrapper';
 import ImageGallery from './ImageGallery/ImageGallery';
 import AddToCart from './AddToCart/AddToCart';
 import ProductInformation from './ProductInformation/ProductInformation';
 import StyleSelector from './StyleSelector/StyleSelector';
 import ProductTextOverview from './ProductTextOverview/ProductTextOverview';
 import ProductIdContext from '../context/ProductIdContext';
+import StyleDataContext from './context/StyleDataContext';
 import SelectedStyleContext from './context/SelectedStyleContext';
 
 const Overview = () => {
   const { productId } = useContext(ProductIdContext);
-  const [selectedStyleIndex, setSelectedStyleIndex] = useState(0);
+  const [styleData, setStyleData] = useState([{}]);
+  const styleDataProvided = { styleData, setStyleData };
+  const [selectedStyleIndex, setSelectedStyleIndex] = useState([]);
   const styleProvidedValue = { selectedStyleIndex, setSelectedStyleIndex };
+
   const ProductInformationWithDetails = withProductDetails(
     ProductInformation,
     productId
@@ -23,22 +27,46 @@ const Overview = () => {
     productId
   );
 
-  const StyleSelectorWithDetails = withStyleDetails(StyleSelector, productId);
+  const findDefaultStyleIndex = (someStyles) => {
+    let result = 0;
+    someStyles.forEach((style, index) => {
+      if (style['default?']) result = index;
+    });
+    return result;
+  };
+
+  useEffect(() => {
+    axios
+      .get('/api/overview/styles/', {
+        params: {
+          productId: productId,
+        },
+      })
+      .then((response) => {
+        setStyleData(response.data.results);
+        setSelectedStyleIndex(findDefaultStyleIndex(response.data.results));
+      })
+      .catch((error) => {
+        console.log('Error getting style details', error);
+      });
+  }, []);
 
   return (
     <div className={styles.container}>
       Product Overview
       <div className={styles.context}>
-        <SelectedStyleContext.Provider value={styleProvidedValue}>
-          <div className={styles.upper}>
-            <ImageGallery />
-            <div className={styles.right}>
-              <ProductInformationWithDetails />
-              <StyleSelectorWithDetails />
-              <AddToCart />
+        <StyleDataContext.Provider value={styleDataProvided}>
+          <SelectedStyleContext.Provider value={styleProvidedValue}>
+            <div className={styles.upper}>
+              <ImageGallery />
+              <div className={styles.right}>
+                <ProductInformationWithDetails />
+                <StyleSelector />
+                <AddToCart />
+              </div>
             </div>
-          </div>
-        </SelectedStyleContext.Provider>
+          </SelectedStyleContext.Provider>
+        </StyleDataContext.Provider>
         <ProductTextOverviewWithDetails />
       </div>
     </div>
