+++ src/components/overview/StyleSelector/StyleSelector.jsx
@@ -1,39 +1,28 @@
 import React, { useContext } from 'react';
-import PropTypes from 'prop-types';
 import styles from './StyleSelector.css';
 import StylePicker from './StylePicker/StylePicker';
 import selectedStyleContext from '../context/SelectedStyleContext';
+import styleDataContext from '../context/StyleDataContext';
 
 // eslint-disable-next-line react/prop-types
-const StyleSelector = ({ styleData }) => {
+const StyleSelector = () => {
+  const { styleData } = useContext(styleDataContext);
   const { selectedStyleIndex } = useContext(selectedStyleContext);
-
+  console.log(styleData, ' index ', selectedStyleIndex);
   return (
     <div className={styles.container}>
       Style Selector
       <div className={styles.title}>
         {'Style > '}
-        <div>{console.log(styleData[selectedStyleIndex], ' index ', selectedStyleIndex) || styleData[selectedStyleIndex].name}</div>
+        <div>
+          {styleData[selectedStyleIndex] === undefined
+            ? 'Placeholder Style'
+            : styleData[selectedStyleIndex].name}
+        </div>
       </div>
       <StylePicker styleData={styleData} />
     </div>
   );
-};
-
-StyleSelector.propTypes = {
-  styleData: PropTypes.arrayOf(
-    PropTypes.shape({
-      name: PropTypes.string,
-    })
-  ),
-};
-
-StyleSelector.defaultProps = {
-  styleData: [
-    {
-      name: 'Placeholder Style Name',
-    },
-  ],
 };
 
 export default StyleSelector;
