+++ src/components/overview/context/StyleDataContext.jsx
@@ -1,0 +1,8 @@
+import { createContext } from 'react';
+
+const StyleDataContext = createContext({
+  styleData: [{}],
+  setStyleData: () => {},
+});
+
+export default StyleDataContext;
